requirements.txt
google-genai==1.61.0
dotenv==0.9.9
streamlit==1.53.1
pypdf==6.6.2
supabase==2.27.2
langchain-text-splitters==1.1.0